updating references
