+++ pylib/packaging/__init__.py
@@ -1,9 +1,12 @@
 
 from .deployment_type import DeploymentType
-from .manifest import ManifestInfo, OfferInfo
+from .manifest import ManifestInfo, OfferInfo, ManifestFile
+from .installer_package import InstallerPackage
 
 __all__ = [
   "ManifestInfo",
   "DeploymentType",
-  "OfferInfo"
+  "OfferInfo",
+  "ManifestFile",
+  "InstallerPackage"
 ]
