+++ pylib/packaging/application_package_builder.py
@@ -1,5 +1,3 @@
-
-
 from zipfile import ZipFile
